File renaming and little bit of clarity
gravity moved to forces/core as one of the 4 fundamental forces
mechanics -> newton_laws as well in core
It is not part of the fundamental forces however, this is because of Rust design tradeoff I tried tweakarounds to dodge complex circular dependencies and enforces clean architecture yet this is a tradeoff and the first bad point of Rust I've seen.
Probably more clarity for the project will have to come because of this meaning files and folders may needs to move around, I'll look into it tomorrow.