Adding Behavior to Utility
Actually, I'm not sure about that one. I don't see LP using classic FSM, nor BT, nor even GOAP (initial consideration for AIs) the only thing is that I'm afraid regarding behavior actions, they'll be coming from liek memory, personality and relationships along trackers + utilityscore to adapt the behavior but to define properly the behavior itself?
Lots of consideration, I'll be playing around with it a little and based on this and the output I'll see if we continue going that way.